requirements.txt
TensorFlow 
Flask
Streamlit
NumPy
Pandas
OpenCV
Pillow
Scikit-learn
Scikit-image
Matplotlib
Seaborn